% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addLegs.R
\name{addLegs}
\alias{addLegs}
\title{Add Legs}
\usage{
addLegs(x = 0, y = 0, xs = 1/100, ys = 1/100, legs = "default",
  gender = c("male", "female"), lwd = 1, linecol = 1, clcol = NULL,
  tail = c(50, 35), w = 5)
}
\arguments{
\item{x}{left bottom alignment of figure}

\item{y}{left bottom alignment of figure}

\item{xs}{x scale (default 1/100)}

\item{ys}{y scale (default 1/100)}

\item{legs}{single character (default \code{'default'})}

\item{gender}{"male", "female"}

\item{lwd}{line weight}

\item{linecol}{color of lines}

\item{clcol}{color of clothes, or NULL to supress clothes (default \code{NULL})}

\item{tail}{x-y coordiates of leg join with torso (default \code{c(50, 35)})}

\item{w}{width parameter (default 5)}
}
\value{
matrix of legs coordinates
}
\description{
Add legs to a plotting area.
Trousers or a skirt/kilt will be added on top of the legs if clcol is specified.
Argument legs can be a character string, or a matrix specifying the
coordinates of the legs.
}
\examples{
plot(0:1, 0:1, type = "n")
addLegs()
}

