% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawStick.R
\name{drawStick}
\alias{drawStick}
\title{Stick Man/Woman}
\usage{
drawStick(x = 0, y = 0, xscale = 1, yscale = 1, gender = c("male",
  "female"), lwd = 3, linecol = 1, hatcol = 2, shcol = NULL,
  clcol = NULL, arms = "default", face = "default", legs = "default",
  hat = "default", ...)
}
\arguments{
\item{x}{left bottom alignment of figure}

\item{y}{left bottom alignment of figure}

\item{xscale}{size of figure}

\item{yscale}{size of figure}

\item{gender}{"male", "female"}

\item{lwd}{line weight}

\item{linecol}{color of lines}

\item{hatcol}{color of hat, or NULL to supress shirt (default \code{NULL})}

\item{shcol}{color of shirt, or NULL to supress shirt (default \code{NULL})}

\item{clcol}{color of clothes, or NULL to supress clothes (default \code{NULL})}

\item{arms}{single character "down", "neutral", "up", "hip", "wave"}

\item{face}{single character "default" ("neutral"), "happy", "sad", "annoyed", "surprised"
    face can also be a numeric matrix with 4 rows and four columns.
    The rows give coordinates of: \enumerate{
    \item left eye x, y, x diameter, y diameter
    \item right eye x, y, x diameter, y diameter
    \item mouth x, y, x diameter, y diameter.
        If the third value of the fourth column is Inf, x, y, x, y of a straight line.
    \item mouth start and stop positions. If the third value is Inf,
        plot a line for the mouth (plus one unused values)
}}

\item{legs}{single character "default" or "apart"}

\item{hat:}{single logical plot hat
or character "none", "shapka", "beanie" or "default" (default "none")}
}
\value{
list with locations of head, arms and legs
}
\description{
Create a stick man or woman.
The stick person is constructed by \code{\link{addArms}}, \code{\link{addLegs}},
and \code{\link{addHead}}.
}
\examples{
# Solitary annoyed figure
   drawStick(face = "surprised", arms = "wave")

   # Map of stick figures
   par(mar = rep(0, 4))

   plot(c(.25,1.25), c(0,3), type = "n", xaxt = 'n', yaxt = 'n', ann = FALSE)

   drawStick(0, 0, arms = "hip")
   drawStick(0.5, 0, gender = "female", arms = "up")

   drawStick(0, 1, arms = "neutral", lwd = 2, linecol = gray(.5),
            clcol = "red", face = "sad")

   drawStick(0.5, 1, gender = "female", arms = "down", clcol = "purple",
            lwd = 2, linecol = gray(.5), face = "sad",hat = TRUE)

   drawStick(0, 2, arms = "wave", linecol = gray(.7),
            clcol = "blue", face = "surprised",hat = FALSE)

   drawStick(0.5, 2, gender = "female", arms = "hip", clcol = "light blue",
            linecol = gray(.7),face = "annoyed")
}
\author{
Francis Smart [cre, cph, aut], Chris Campbell (Mango Solutions) [aut]
}

