% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotStick.R
\name{plotStick}
\alias{plotStick}
\title{Plot Stick Men and Women at coordinates}
\usage{
plotStick(x, y, xlim, ylim, xlab = "x", ylab = "y", lwd = 1,
  linecol = 1, hatcol = NULL, shcol = NULL, clcol = NULL, col = NULL,
  gender = c("male", "female"), arms = "default", face = "default",
  legs = "default", hat = "default", tower = 10, newplot = TRUE, ...)
}
\arguments{
\item{x}{numeric vector, progress of stick people from left to right}

\item{y}{numeric vector, progress of stick people from bottom to top}

\item{xlim}{length 2 numeric, x-axis limits to use, otherwise range of x expanded}

\item{ylim}{length 2 numeric, y-axis limits to use, otherwise range of y expanded}

\item{xlab}{single character, x axis label}

\item{ylab}{single character, y axis label}

\item{lwd}{line weight}

\item{linecol}{color of lines}

\item{hatcol}{color of hat, or NULL to supress hat (default \code{NULL})}

\item{shcol}{color of shirt, or NULL to supress shirt (default \code{NULL})}

\item{clcol}{color of clothes, or NULL to supress clothes (default \code{NULL})}

\item{col}{matrix of colors for fill, overriding shcol, clcol and hatcol
with up to three columns (default \code{NULL})}

\item{gender}{character vector "male", "female"}

\item{arms}{character vector "down", "neutral", "up", "hip", "wave"}

\item{face}{list or character vector "default" ("neutral"), "happy", "sad", "annoyed", "surprised"
face can also be a numeric matrix with 4 rows and four columns.
See \code{\link{drawStick}} for more details.}

\item{legs}{single character "default" or "apart"}

\item{hat}{single logical plot hat or automatic if NA (default NA)}

\item{tower}{single numeric approximate number of stick people
that could be stacked in the plotting area (default 10)}

\item{newplot}{single logical should new plot be created before adding stick men? (default TRUE)}
}
\description{
Create a stick person plot.
A new plot is created with points specified by x and y.
Stick people are scaled so that ten people could be stacked
head to toe in the plotting area. Set argument tower to change this.
Stick people will not wear clothes unless col is specified.
Their line colours can be specified with linecol.
}
\examples{
plotStick(x = 1:10, y = c(1:5, 5:1))
points(1:10, c(1:5, 5:1), col = 2)
points(1:10, c(1:5, 5:1), col = 2, pch = 3)
plotStick(x = 1:10, y = sin(1:10), hatcol = 2, shcol = rainbow(10),
    clcol = 1:10, gender = c("male", "female"),
    arms = c("down", "neutral", "up", "hip", "wave"))
plotStick(x = 1:10, y = cos(1:10),
    hat = c("none", "shapka", "beanie", "fedora"), col = rainbow(30), cex = 2)
}
\author{
Francis Smart [cre, cph, aut], Chris Campbell (Mango Solutions) [aut]
}

