% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pointsStick.R
\name{pointsStick}
\alias{pointsStick}
\title{Add Stick Men and Women at coordinates}
\usage{
pointsStick(x, y, lwd = 1, linecol = 1, hatcol = NULL, shcol = NULL,
  clcol = NULL, col = NULL, gender = c("male", "female"),
  arms = "default", face = "default", legs = "default", hat = "default",
  tower = 10)
}
\arguments{
\item{x}{numeric vector, progress of stick people from left to right}

\item{y}{numeric vector, progress of stick people from bottom to top}

\item{lwd}{line weight}

\item{linecol}{color of lines}

\item{hatcol}{color of hat, or NULL to supress hat (default \code{NULL})}

\item{shcol}{color of shirt, or NULL to supress shirt (default \code{NULL})}

\item{clcol}{color of clothes, or NULL to supress clothes (default \code{NULL})}

\item{col}{matrix of colors for fill, overriding shcol, clcol and hatcol
with up to three columns (default \code{NULL})}

\item{gender}{character vector "male", "female"}

\item{arms}{character vector "down", "neutral", "up", "hip", "wave"}

\item{face}{list or character vector "default" ("neutral"), "happy", "sad", "annoyed", "surprised"
face can also be a numeric matrix with 4 rows and four columns.
See \code{\link{drawStick}} for more details.}

\item{legs}{single character "default" or "apart"}

\item{hat}{single logical plot hat or automatic if NA (default NA)}

\item{tower}{single numeric approximate number of stick people
that could be stacked in the plotting area (default 10)}
}
\description{
Add Stick Men or Women to a plot.
This function is a convenience wrapper to \code{\link{plotStick}}
with newplot FALSE.
}
\examples{
plot(x = 1:10, y = c(1:5, 5:1), cex = 8)
pointsStick(1:10, c(1:5, 5:1), col = 2)
plot(x = 1:10, y = sin(1:10), type = "n")
pointsStick(x = 1:10, y = sin(1:10), hatcol = 2, shcol = rainbow(10),
    clcol = 1:10, gender = c("male", "female"),
    arms = c("down", "neutral", "up", "hip", "wave"))
}
\author{
Francis Smart [cre, cph, aut], Chris Campbell (Mango Solutions) [aut]
}

