% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawStick.R
\docType{package}
\name{stick-package}
\alias{stick}
\alias{stick-package}
\title{Craft Stick Men and Women}
\description{
Make plots with stick people using \code{\link{plotStick}}.
Add stick people to an existing core graphics plot using \code{\link{pointsStick}}.
}
\details{
Run tests using \pkg{testthat} call \code{test_package("stick")}.
Comparison images can be checked using
\pkg{\href{https://github.com/MangoTheCat/visualTest}{visualTest}}.

\tabular{ll}{
Package: \tab stick \cr
Type: \tab Package \cr
Version: \tab 1.0.12 \cr
Date: \tab Oct 2015 \cr
Lazyload: \tab yes \cr
}
}
\author{
Francis Smart [cre, cph, aut], Chris Campbell (Mango Solutions) [aut]
}
\references{
\href{http://www.econometricsbysimulation.com/2014/01/stick-figure-function-r.html}{www.econometricsbysimulation.com (stick figures)}
}
\keyword{package}

