\docType{package}
\name{stick-package}
\alias{stick}
\alias{stick-package}
\title{Craft Stick Men and Women}
\description{
Make Stick People. Run tests using \pkg{testthat} call
\code{test_package("stick")}.
}
\details{
\tabular{ll}{ Package: \tab stick \cr Type: \tab Package
\cr Version: \tab 1.0.11 \cr Date: \tab May 2014 \cr
Lazyload: \tab yes \cr }
}
\author{
Francis Smart, Mango Solutions
}
\references{
\url{http://www.econometricsbysimulation.com/2014/01/stick-figure-function-r.html}
}
\keyword{package}

