\docType{package}
\name{stick-package}
\alias{stick}
\alias{stick-package}
\title{Craft Stick Men and Women}
\description{
Make Stick People
}
\details{
\tabular{ll}{ Package: \tab stick \cr Type: \tab Package
\cr Version: \tab 1.0.9 \cr Date: \tab February 2014 \cr
Lazyload: \tab yes \cr }
}
\author{
Francis Smart, Mango Solutions
}
\keyword{package}

